Implement on the fly rewriting of URLs in contents
A very simplistic hack to allow rewriting URLs in contents we are importing. With this change the --rewrite-host flag allow us to replace all occurrences of "http://localhost" with the value of the host param during import.

This allows us to write sets of data containing references to localhost in any field but import them as a reference to the host they are getting written to.