Fix mime-type sending
Graviton since to have a recursion where sniffing the mime-type does not work on the server side and for some reason guzzle does not pick up the right mime type when it only gets the files contents. I'm fixing this here since we need to get this done and I don't see us having the luxury of enough time to look into what really going on at the server side.